Move annotations to stage package.
